requirements.txt
Django==3.2.16
Pillow==10.4.0
psycopg2-binary==2.9.12
python-dotenv==1.2.2
